+++ app/models/addon-operation-invocation.ts
@@ -1,8 +1,21 @@
 import Model, { AsyncBelongsTo, attr, belongsTo } from '@ember-data/model';
 
 import UserReferenceModel from 'ember-osf-web/models/user-reference';
-import ConfiguredAddonModel, { ConnectedOperationNames } from 'ember-osf-web/models/configured-addon';
+import ConfiguredAddonModel from 'ember-osf-web/models/configured-addon';
 import AuthorizedAccountModel from 'ember-osf-web/models/authorized-account';
+
+export enum ConnectedOperationNames {
+    HasRevisions = 'has_revisions',
+    ListRootItems = 'list_root_items',
+    ListChildItems = 'list_child_items',
+    GetItemInfo = 'get_item_info',
+}
+
+export interface OperationKwargs {
+    itemId?: string;
+    itemType?: ItemType;
+    pageCursor?: string;
+}
 
 export enum InvocationStatus {
     STARTING = 'STARTING',
@@ -38,7 +51,7 @@ export interface Item {
 export default class AddonOperationInvocationModel extends Model {
     @attr('string') invocationStatus!: InvocationStatus;
     @attr('string') operationName!: ConnectedOperationNames;
-    @attr('object', {snakifyForApi: true}) operationKwargs!: any;
+    @attr('object', {snakifyForApi: true}) operationKwargs!: Partial<OperationKwargs>;
     @attr('object', {snakifyForApi: true}) operationResult!: OperationResult;
     @attr('date') created!: Date;
     @attr('date') modified!: Date;
