+++ app/models/authorized-account.ts
@@ -1,4 +1,12 @@
 import Model, { attr } from '@ember-data/model';
+import { waitFor } from '@ember/test-waiters';
+import { task } from 'ember-concurrency';
+import { ConnectedOperationNames, OperationKwargs } from 'ember-osf-web/models/addon-operation-invocation';
+
+export enum ConnectedCapabilities {
+    Access = 'ACCESS',
+    Update = 'UPDATE',
+}
 
 export interface AddonCredentialFields {
     username?: string;
@@ -24,4 +32,29 @@ export default class AuthorizedAccountModel extends Model {
     @attr('boolean') initiateOauth!: boolean; // write-only
     @attr('fixstring') readonly authUrl!: string; // Only returned when POSTing to /authorized-xyz-accounts
     @attr('boolean') readonly credentialsAvailable!: boolean;
+
+    @task
+    @waitFor
+    async getFolderItems(this: AuthorizedAccountModel, kwargs?: OperationKwargs) {
+        const operationKwargs = kwargs || {};
+        const operationName = operationKwargs.itemId ? ConnectedOperationNames.ListChildItems :
+            ConnectedOperationNames.ListRootItems;
+        const newInvocation = this.store.createRecord('addon-operation-invocation', {
+            operationName,
+            operationKwargs,
+            thruAccount: this,
+        });
+        return await newInvocation.save();
+    }
+
+    @task
+    @waitFor
+    async getItemInfo(this: AuthorizedAccountModel, itemId: string) {
+        const newInvocation = this.store.createRecord('addon-operation-invocation', {
+            operationName: ConnectedOperationNames.GetItemInfo,
+            operationKwargs: { itemId },
+            thruAccount: this,
+        });
+        return await newInvocation.save();
+    }
 }
